PLAT-42690 fix Table selection behavior on rows update
Motivation:
In connectoradmin, when the user changes pages or filters the list, the state
of the selectedIndices is not updated properly, so the active row is not
highlighted properly in the table.

Consideration:
* none

Changes:
* In the Table's componentDidMount hook, we set the selectedIndices state
according to the same logic that initializes it in the constructor.
* selectedRowIndices variables were renamed to selectedIndices to reflect
TableState type definition.
* cleanFlow script was renamed to cleanflow to match the package.json/other
script names

Testing:
* manual testing of UI
* ran flow
* ran lint
* ran validatecomponents
* ran fullbuild

Impact: [UI]