PLAT-41183: Fix logout
Add a delay before redirecting for logging out since
the new page was being loaded before the old state
was finshed being removed and the page re-rendered
(with no user name) but wasn't really logged in.
This only seemed to happen if the page was manually
refreshed prior to choosing the logout item in the
masthead.